feat(bars): converge the Space Bar focused accent onto the palette
`SpaceBarStyle.focusedItemColor` was the last colour still off the
KiwiCanopy palette — the amber `#E8A33D` #439 deliberately left
alone, since the bar's "viewing-not-active" accent is a different
semantic from the drag drop-zone and the bar owns its own backdrop,
so it never needed the drop-zone's contrast darkening.

It turns out to need the identical darkening for a different
reason, and that reason is what decided the hue.

## The rule the palette was tested against does not hold for green

`docs/design-decisions.md` requires the focused accent to be the
complementary *temperature* of the primary, and
`ColorPaletteTests.focusedAccentDistinct` pins a *different hue*.
Both held for `#8DB354` vs `#E8A33D`, and both are insufficient:
green↔amber is precisely the axis protanopia and deuteranopia
flatten. Simulated, the shipped default pair separates at **22 of
441** under protanopia — it renders `#AFAF54` vs `#ADAD3E`, the
same colour. Every *green*-primary bundled palette collapses the
same way (Kiwi Neon 39, Kiwi Gold 49); every *cool*-primary one
clears 175. The brand's own green-forward identity lands the
default on the one pair the rule cannot protect.

Lightness is the channel every CVD type preserves, so the rule
gains a second clause: different hue **and** a lightness gap.

## What shipped

`#C2790A` — the amber #439 had *already* ratified for the drag
drop-zone. Same hue (H36) as the old default, L57% → L40%. It
converges onto the palette by reusing a hex the rebrand chose
rather than minting a second amber, it is the same "keep the hue,
drop the lightness" move the identity already makes for ink,
borders and the ring, and it lifts the pair's protan separation
from 22 to **93**.

Accepted trade, and the one thing to look at on device: focused
now reads *darker* than the active green rather than brighter.

Kiwi Neon and Kiwi Gold are left as they are. Retuning either is
an eye-confirm call of its own, and a catalog-wide
perceptual-distance assertion would fail both today — so
tightening `focusedAccentDistinct` and retuning them belong to one
change, not this one. `ColorPaletteTests` keeps its inequality
check; the new `SpaceBarAccentSeparationTests` pins the lightness
gap for the derived default palette only, with a floor sized
against measured values (shipped 1.443; pre-#470 1.118; the two
near-miss retunes 1.344 and 1.022).

The issue also asked to update the `Kiwi Dark` bundled palette —
that palette was retired in #439, and Kiwi Gold's
`focused_item_color` is already the green `#8DB354`, so the struct
default was the only remaining holdout.

## Note on the file split

`SpaceBarStyle.swift` sat at 349 of its 350-line ceiling, so the
field's rationale could not be written down without splitting.
Its `Codable` members move to `SpaceBarStyle+Coding.swift`, the
same convention as `TilingSettings+Coding.swift`. The conformance
declaration stays in the type's own file: Swift only synthesizes
`encode(to:)` where the conformance sits, and a cross-file
`extension … : Codable` would have forced a hand-written encode —
exactly the mirrored field list `.claude/rules/parity-tests.md`
exists to prevent.

fixes #470
